corresponding changes in perfherder for moving files to shared folder
